Move utils to util.ts
